# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SecureBank Core Banking API
  description: "Autenticação (JWT, refresh com rotação, MFA TOTP), clientes, contas,\
    \ depósitos, saques, transferências, pagamentos, extrato, limites, auditoria e\
    \ administração. Valores monetários trafegam como string (\"100.00\"). Transferê\
    ncias e pagamentos exigem o header Idempotency-Key."
  version: v1
servers:
- url: http://localhost:8100
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/transfers:
    get:
      tags:
      - transfer-controller
      operationId: list
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferPage"
    post:
      tags:
      - transfer-controller
      operationId: create
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponse"
  /api/v1/security/password:
    post:
      tags:
      - security-controller
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/v1/security/mfa:
    post:
      tags:
      - security-controller
      operationId: setupMfa
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MfaSetupResponse"
    delete:
      tags:
      - security-controller
      operationId: disableMfa
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/v1/security/mfa/confirm:
    post:
      tags:
      - security-controller
      operationId: confirmMfa
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/v1/payments:
    get:
      tags:
      - payment-controller
      operationId: list_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentPage"
    post:
      tags:
      - payment-controller
      operationId: create_1
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/v1/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /api/v1/auth/mfa/verify:
    post:
      tags:
      - auth-controller
      operationId: verifyMfa
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/v1/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /api/v1/transfers/{id}:
    get:
      tags:
      - transfer-controller
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransferResponse"
...
  /api/v1/payments/{id}:
    get:
      tags:
      - payment-controller
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
  /api/v1/security/sessions/{id}:
    delete:
      tags:
      - security-controller
      operationId: revokeSession
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
components:
  schemas:
...
    MoneyResponse:
      type: object
      properties:
        amount:
          type: string
        currency:
          type: string
    TransferRequest:
      type: object
      properties:
        sourceAccountId:
          type: string
          format: uuid
        destinationBranch:
          type: string
          maxLength: 4
          minLength: 0
        destinationAccountNumber:
          type: string
          maxLength: 14
          minLength: 0
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
          maxLength: 140
          minLength: 0
      required:
      - amount
      - destinationAccountNumber
      - destinationBranch
      - sourceAccountId
    TransferResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sourceAccountId:
          type: string
          format: uuid
        destinationAccountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MoneyResponse"
        description:
          type: string
        status:
          type: string
        debitTransactionId:
          type: string
          format: uuid
        creditTransactionId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          maxLength: 128
          minLength: 0
        newPassword:
          type: string
          maxLength: 128
          minLength: 0
      required:
      - currentPassword
      - newPassword
    MfaSetupResponse:
      type: object
      properties:
        secret:
          type: string
        otpauthUri:
          type: string
    CodeRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 6
          minLength: 6
      required:
      - code
    PaymentRequest:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        barcode:
          type: string
          maxLength: 48
          minLength: 0
        description:
          type: string
          maxLength: 140
          minLength: 0
      required:
      - accountId
      - amount
      - barcode
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MoneyResponse"
        barcode:
          type: string
        description:
          type: string
        status:
          type: string
        transactionId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        document:
          type: string
          maxLength: 14
          minLength: 0
        email:
          type: string
          maxLength: 254
          minLength: 0
        phone:
          type: string
          maxLength: 16
          minLength: 0
        password:
          type: string
          maxLength: 128
          minLength: 0
      required:
      - document
      - email
      - name
      - password
      - phone
    RegisterResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        email:
          type: string
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - refreshToken
    TokenResponse:
      type: object
      properties:
        mfaRequired:
          type: boolean
        mfaToken:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
    MfaVerifyRequest:
      type: object
      properties:
        mfaToken:
          type: string
          maxLength: 2000
          minLength: 0
        code:
          type: string
          maxLength: 6
          minLength: 6
      required:
      - code
      - mfaToken
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          minLength: 0
        password:
          type: string
          maxLength: 128
          minLength: 0
      required:
      - email
      - password
...
    TransferPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransferResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
...
    PaymentPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PaymentResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Access token (RS256, 15 min) obtido em POST /api/v1/auth/login"
      scheme: bearer
      bearerFormat: JWT